+++ pub-critic/src/app/api/ReviewsApi.ts
@@ -1,36 +1,6 @@
 import axios from "axios";
 import { baseURL } from "./Shared";
-
-export interface PostReviewProps {
-  title: string;
-  body: string;
-  score: number;
-  gameName: string;
-}
-
-export interface Author {
-  id: number;
-  name: string;
-}
-
-export interface Avarage {
-  gameId: number;
-  count: number;
-  avarage: number;
-}
-
-export interface Review {
-  id: number;
-  title: string;
-  body: string;
-  score: number;
-  gameId: number;
-  userId: number;
-  gameName: string;
-  author: Author;
-  date: string;
-}
-
+import { PostReviewProps } from "common/interfaces";
 const api = axios.create({
   baseURL: baseURL,
   headers: {
