+++ pub-critic/src/app/components/Review/Review.module.scss
@@ -22,6 +22,7 @@
     display: flex;
     flex-direction: row;
     align-items: center;
+    
     .image {
       width: vw-calc(80);
       height: vw-calc(80);
